[llvm/CAS] Refactor `OnDiskCAS` into 2 classes, `OnDiskGraphDB` and `OnDiskCAS`
`OnDiskGraphDB` represents the underlying storage implementation, is hash-algorithm independent, and decoupled from the `ObjectStore` API.
`OnDiskCAS` is implemented as a wrapper of an `OnDiskGraphDB` instance.

Essentially the provided abstractions are like this:
`OnDiskHashMapTrie` + CAS data storage -> `OnDiskGraphDB` + `BuiltinObjectHasher` -> `OnDiskCAS`

`OnDiskGraphDB` is useful as a building block that can be combined with any hash-algorithm for implementing an on-disk CAS.
Since it is decoupled from the `ObjectStore` API it can have additional flexibility that may be useful to clients;
specifically `OnDiskGraphDB`'s object references can be created by the hash alone without requiring that references point to stored objects.

Other notable differences are some API name differences to improve clarity and using iterators for reading object references.
